requirements.txt
python-dotenv>=1.0.0
groq>=0.9.0
requests>=2.31.0
scikit-learn>=1.3.0
pillow>=10.0.0
